devtools/client/shared/components/reps/test/mochitest/test_reps_undefined.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test undefined rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - undefined</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
window.onload = Task.async(function* () {
  try {
    let ReactDOM = browserRequire("devtools/client/shared/vendor/react-dom");
    let React = browserRequire("devtools/client/shared/vendor/react");
    const {
      REPS,
      getRep,
    } = browserRequire("devtools/client/shared/components/reps/reps");
    let { Undefined } = REPS;

    let gripStub = {
      "type": "undefined"
    };

    // Test that correct rep is chosen
    is(getRep(gripStub), Undefined.rep, "Rep correctly selects Undefined Rep");

    // Test rendering
    const renderedComponent = renderComponent(Undefined.rep, {});
    is(renderedComponent.className, "objectBox objectBox-undefined", "Undefined rep has expected class names");
    is(renderedComponent.textContent, "undefined", "Undefined rep has expected text content");
  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
});
</script>
</pre>
</body>
</html>
